:recycle: refactor: 주문 생성 시 재고 차감 로직 삭제 및 주문 수정의 메서드 간결화
