+++ src/main/java/com/example/evostyle/domain/orderitem/repository/OrderItemsRepository.java
@@ -1,6 +1,6 @@
 package com.example.evostyle.domain.orderitem.repository;
 
-import com.example.evostyle.domain.orderitem.entity.OrderItem;
+import com.example.evostyle.domain.order.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderItemsRepository extends JpaRepository<OrderItem,Long> {
